+++ src/encoding.h
@@ -4,8 +4,8 @@
 extern Instruction emptyInstruction;
 
 // Use somente para valores sem sinal (registradores rs, rt e rd, etc)
-unsigned int binaryToUnsignedInt(const char *serializedBinary);
+uint8_t binaryToUnsignedInt(const char *serializedBinary);
 
 void extendSignal(char *baseSerializedBinary, char *buffer, int size);
 
-int complementOfTwoToInt(const char *serializedBinary);
+int8_t complementOfTwoToInt(const char *serializedBinary);
